PDI-10310 : Fix all -Xlint:all warnings in PDI
git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/trunk@18790 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ engine/src/org/pentaho/di/job/entries/syslog/SyslogDefs.java
@@ -121,7 +121,7 @@ public class SyslogDefs {
 	}
 
 	static public int getPriority( String priority )throws SyslogException {
-		Integer result = (Integer) SyslogDefs.priHash.get( priority );
+		Integer result = SyslogDefs.priHash.get( priority );
 		
 		if ( result == null ){
 			throw new SyslogException (BaseMessages.getString(PKG, "JobEntrySyslog.UnknownPriotity", priority));
@@ -131,7 +131,7 @@ public class SyslogDefs {
 	}
 
 	static public int getFacility( String facility ) throws SyslogException {
-		Integer result = (Integer) SyslogDefs.facHash.get( facility );
+		Integer result = SyslogDefs.facHash.get( facility );
 
 		if ( result == null ) {
 			throw new SyslogException (BaseMessages.getString(PKG, "JobEntrySyslog.UnknownFacility", facility));
